split and make SmachViewerBase, utils and text_wrapper
